+++ src/operators/video_fade.ml
@@ -22,8 +22,6 @@
 
 open Source
 
-module Img = Image.RGBA32
-
 (** Fade-in at the beginning of every track.
   * The [duration] is in seconds. *)
 class fade_in ~kind ?(meta="liq_video_fade_in") duration
@@ -76,7 +74,7 @@ object
               if count < length then
                 for i=0 to min (len-1) (length-count-1) do
                   let m = fade (count+i) in
-                    fadefun rgb.(off+i) m
+                    fadefun (Video.get rgb (off+i)) m
                 done ;
               if state <> `Idle then
                 state <- `Play (fade,fadefun,length,count+len)
@@ -136,7 +134,7 @@ object
                   let rgb = rgb.(0) in
                     for i=0 to len-1 do
                       let m = fade (n-i) in
-                        fadefun rgb.(off+i) m
+                        fadefun (Video.get rgb (off+i)) m
                     done
               | None -> ()
             end 
@@ -164,31 +162,45 @@ let proto =
   ]
 
 let rec transition_of_string p transition =
+  let translate img x y =
+    let tmp = Video.Image.copy img in
+    Image.YUV420.fill_alpha img 0;
+    Video.Image.add tmp ~x ~y img
+  in
   let ifm n a = int_of_float ((float_of_int n) *. a) in
     match transition with
-      | "fade" -> fun () -> Img.Effect.Alpha.scale
+      | "fade" ->
+          fun () img t ->
+          Video.Image.fill_alpha img (ifm 256 t)
       | "slide_left" ->
           fun () buf t ->
-            Img.Effect.translate buf
-              (ifm (Lazy.force Frame.video_width) (t-.1.)) 0
+            translate buf (ifm (Lazy.force Frame.video_width) (t-.1.)) 0
       | "slide_right" ->
           fun () buf t ->
-            Img.Effect.translate buf
-              (ifm (Lazy.force Frame.video_width) (1.-.t)) 0
+            translate buf (ifm (Lazy.force Frame.video_width) (1.-.t)) 0
       | "slide_up" ->
           fun () buf t ->
-            Img.Effect.translate buf
-              0 (ifm (Lazy.force Frame.video_height) (1.-.t))
+            translate buf 0 (ifm (Lazy.force Frame.video_height) (1.-.t))
       | "slide_down" ->
           fun () buf t ->
-            Img.Effect.translate buf
-              0 (ifm (Lazy.force Frame.video_height) (t-.1.))
-      | "grow" -> fun () buf t -> Img.Effect.affine buf t t 0 0
+            translate buf 0 (ifm (Lazy.force Frame.video_height) (t-.1.))
+      | "grow" ->
+         fun () img t ->
+         let w = Video.Image.width img in
+         let h = Video.Image.height img in
+         let w' = ifm w t in
+         let h' = ifm h t in
+         let tmp = Video.Image.create w' h' in
+         Video.Image.scale img tmp;
+         Video.Image.fill_alpha img 0;
+         let x = (w - w') / 2 in
+         let y = (h - h') / 2 in
+         Video.Image.add ~x ~y tmp img
       | "disc" ->
           let w = Lazy.force Frame.video_width in
           let h = Lazy.force Frame.video_height in
           let r_max = int_of_float (sqrt (float_of_int (w * w + h * h))) / 2 in
-            fun () buf t -> Img.Effect.Alpha.disk buf (w/2) (h/2) (ifm r_max t)
+          fun () buf t -> Image.YUV420.disk_alpha buf (w/2) (h/2) (ifm r_max t)
       | "random" ->
           let trans =
             [|"slide_left"; "slide_right"; "slide_up"; "slide_down";