+++ src/operators/video_text.ml
@@ -22,8 +22,6 @@
 
 open Source
 
-module Img = Image.RGBA32
-
 class text ~kind
   init render_text ttf ttf_size color tx ty speed cycle meta text (source:source) =
   let () = init () in
@@ -49,13 +47,13 @@ object (self)
   val mutable cur_text = text ()
 
   method private render_text text =
-    let w, h, get_pixel = render_text ~font:ttf ~size:ttf_size text in
-    let tf = Img.create w h in
+    let w, h, get_pixel_rgba = render_text ~font:ttf ~size:ttf_size text in
+    let tf = Video.Image.create w h in
     let tr, tg, tb = Image.RGB8.Color.of_int color in
     for y = 0 to h - 1 do
       for x = 0 to w - 1 do
-        let a = get_pixel x y in
-        Img.set_pixel tf x y (tr, tg, tb, a)
+        let a = get_pixel_rgba x y in
+        Video.Image.set_pixel_rgba tf x y (tr, tg, tb, a)
       done
     done;
     text_frame <- Some tf
@@ -73,7 +71,7 @@ object (self)
     | Some (rgb,off,len) ->
        let rgb = rgb.(0) in
        let tf = self#get_text_frame in
-       let tfw = Img.width tf in
+       let tfw = Video.Image.width tf in
        let text =
          match meta with
          | None -> text ()
@@ -96,13 +94,13 @@ object (self)
        for i = off to off + len - 1 do
          if speed = 0 then
            (
-             Img.add tf rgb.(i) ~x:pos_x ~y:pos_y ;
+             Video.Image.add tf (Video.get rgb i) ~x:pos_x ~y:pos_y ;
              pos_x <- tx () ;
              pos_y <- ty ()
            )
          else
            (
-             if pos_x <> -tfw then Img.add tf rgb.(i) ~x:pos_x ~y:pos_y ;
+             if pos_x <> -tfw then Video.Image.add tf (Video.get rgb i) ~x:pos_x ~y:pos_y ;
              pos_x <- pos_x - speed ;
              if pos_x < -tfw then
                if cycle then