+++ src/outputs/graphics_out.ml
@@ -44,7 +44,7 @@ object
         c.Frame.video.(0)
     in
       for frame = 0 to 0 do
-        let img = Image.RGBA32.to_int_image rgb.(frame) in
+        let img = Video.Image.to_int_image (Video.get rgb frame) in
         let img = Graphics.make_image img in
           Graphics.draw_image img 0 0
       done