+++ CHANGES.md
@@ -1,4 +1,23 @@
-1.4.0 ()
+1.5.0 (unreleased)
+=====
+
+New:
+
+- Switch to YUV420 as internal image format, much more efficient (#848).
+- Added `output.file.dash.ffmpeg`.
+- Added LV2 support (#906).
+- Added `string.nth` (#970).
+- Added `string.binary.to_int` (#970).
+
+Changed:
+
+- Native Liquidsoap implementation of list functions (#920).
+
+Fixed:
+
+- Fix implementation of recursive functions (#934).
+
+1.4.0 (29-09-2019)
 =====
 
 New:
@@ -8,6 +27,8 @@ New:
 - Added support for deconstructing tuples: `let (z,t,_) = x` (#838)
 - Added `input.{file,harbor}.hls` to read HLS stream (#59, #295, #296).
 - Added `output.hls` to natively stream in HLS (#758).
+- Added `%ffmpeg` native encoder, only for audio encoding for now (#952)
+- Added ffmpeg-based stream decoder, limited to mime type `application/ffmpeg` for now.
 - Added `(to_){string,float,int,bool}_getter` operators to handle getters in
   script side.
 - Made `p` parameter in `smooth_add` a `float` getter (#601)
@@ -37,7 +58,6 @@ New:
 - Added support for bash completion.
 - Added `video.add_text.native`.
 - Added `configure.bindir`
-- Added ffmpeg-based stream decoder, limited to mime type `application/ffmpeg` for now.
 - Added `for` and `while` loop functions.
 - Added `list.case`.
 - Added `metadata.string_getter` and `metadata.float_getter`.
@@ -47,6 +67,7 @@ New:
 - Added `path.remove_extension`.
 - Added SSL read/write timeout options, use it for incoming socket connections (#932)
 - Added ffmpeg resampler (#947).
+- Added `lsl` and `lsr`.
 
 Changed:
 
@@ -112,6 +133,8 @@ Changed:
 - Added `phase_inversion={true/false}` to `%opus` encoder (#937)
 - Fixed encoders forcing frame rate and audio channels too early (#933)
 - Change filename to a string getter in file-based outputs. (#198)
+- Changed `audio.converter.samplerate.preferred` option to
+  `audio.converter.samplerate.converters` to give a list of possible converters.
 
 Fixed: